requirements.txt
Django==4.1.7
django-colorfield==0.8.0
djangorestframework==3.14.0
djoser==2.1.0
sqlparse~=0.4.3